union of supplier payment and information and resolve entity make migration file
