+++ assets/styles/clientPayment.scss
@@ -7,6 +7,7 @@
 
         table {
             text-align: center;
+            margin: 0 auto;
             width: 90%;
             max-width: 40rem;
             font-style: $Helvetica;
