--- assets/styles/supplierInformation.scss
@@ -1,79 +1,0 @@
-@import "variables.scss";
-
-.supplierInformation {
-    .informations {
-        background-color: $secondBackground;
-        padding-bottom: 0.75rem;
-
-        table {
-            margin: 0 auto;
-            text-align: center;
-            width: 90%;
-            max-width: 40rem;
-            font-style: $Helvetica;
-            border-collapse: separate;
-            border-spacing: 0 0.25rem;
-
-            thead {
-                font-size: 0.938rem;
-                border: none;
-
-                th {
-                    vertical-align: middle;
-                }
-            }
-
-            tbody {
-                background-color: $thirdBackground;
-                border: none;
-                font-size: 0.813rem;
-
-                td {
-                    height: 2.5rem;
-                    border-spacing: 0.25rem;
-                }
-
-                td:first-child {
-                    border-radius: 5px 0 0 5px;
-                }
-
-                td:last-child {
-                    border-radius: 0 5px 5px 0;
-                }
-            }
-        }
-    }
-
-    form {
-        margin: 0 auto;
-        width: 85%;
-        max-width: 22rem;
-
-        label {
-            font-family: $Helvetica;
-            font-size: 1rem;
-            color: black;
-            margin-top: 1.25rem;
-            margin-bottom: 0.375rem;
-        }
-
-        input {
-            height: 2.25rem;
-        }
-
-        textarea {
-            height: 6.25rem;
-        }
-
-        .btn {
-            display: block;
-            margin: 1rem auto;
-            text-align: center;
-            width: 9.625rem;
-            height: 2.25rem;
-            text-align: center;
-            background-color: $green;
-            color: $light;
-        }
-    }
-}
