+++ src/Controller/SupplierController.php
@@ -40,7 +40,7 @@ class SupplierController extends AbstractController
             $entityManager->persist($supplier);
             $entityManager->flush();
 
-            return $this->redirectToRoute('supplier_information_new', [
+            return $this->redirectToRoute('supplier_payment_new', [
                 'booking_id' => $booking->getId()
             ], Response::HTTP_SEE_OTHER);
         }
