+++ src/Entity/SupplierPayment.php
@@ -15,12 +15,47 @@ class SupplierPayment
     public const STATUS = ['due', 'paid'];
     public const TYPES = ['deposit', 'final_payment', 'fulll_payment'];
     public const MODES = ['credit_card', 'wire_transfert', 'check', 'credit', 'refund'];
+    public const CURRENCY = [
+        'EUR', 'GBP', 'BGN', 'HRK', 'DKK', 'HUF', 'PLN', 'SEK', 'CZK', 'RON', 'ALL',
+        'BYN', 'BAM', 'ISK', 'CHF', 'MKD', 'MDL', 'NOK', 'RSD', 'UAH', 'GIP',
+    ];
+
     /**
      * @ORM\Id
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
      */
     private int $id;
+
+        /**
+     * @ORM\Column(type="float", nullable=true)
+     * @Assert\Positive
+     */
+    private ?float $dueAmount;
+
+    /**
+     * @ORM\Column(type="date", nullable=true)
+     */
+    private ?DateTimeInterface $dueDate;
+
+    /**
+     * @ORM\Column(type="float", nullable=true)
+     * @Assert\Positive
+     */
+    private ?float $exchangeRate;
+
+    /**
+     * @ORM\Column(type="float", nullable=true)
+     * @Assert\Positive
+     */
+    private ?float $dueDollarsAmount;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     * @Assert\Choice(choices=SupplierPayment::CURRENCY, message="Choose a valid currency.")
+     */
+    private ?string $currency;
+
 
     /**
      * @ORM\Column(type="date", nullable=true)
@@ -200,6 +235,66 @@ class SupplierPayment
     public function setMode(?string $mode): self
     {
         $this->mode = $mode;
+
+        return $this;
+    }
+
+    public function getDueAmount(): ?float
+    {
+        return $this->dueAmount;
+    }
+
+    public function setDueAmount(?float $dueAmount): self
+    {
+        $this->dueAmount = $dueAmount;
+
+        return $this;
+    }
+
+    public function getDueDate(): ?\DateTimeInterface
+    {
+        return $this->dueDate;
+    }
+
+    public function setDueDate(?\DateTimeInterface $dueDate): self
+    {
+        $this->dueDate = $dueDate;
+
+        return $this;
+    }
+
+    public function getExchangeRate(): ?float
+    {
+        return $this->exchangeRate;
+    }
+
+    public function setExchangeRate(?float $exchangeRate): self
+    {
+        $this->exchangeRate = $exchangeRate;
+
+        return $this;
+    }
+
+    public function getDueDollarsAmount(): ?float
+    {
+        return $this->dueDollarsAmount;
+    }
+
+    public function setDueDollarsAmount(?float $dueDollarsAmount): self
+    {
+        $this->dueDollarsAmount = $dueDollarsAmount;
+
+        return $this;
+    }
+
+    public function getCurrency(): ?string
+    {
+        return $this->currency;
+    }
+
+    public function setCurrency(?string $currency): self
+    {
+        $this->currency = $currency;
 
         return $this;
     }
