+++ src/Form/SupplierPaymentType.php
@@ -24,6 +24,47 @@ class SupplierPaymentType extends AbstractType
                 'label' => 'Supplier',
                 'placeholder' => 'Choose a supplier',
             ])
+            ->add('currency', ChoiceType::class, [
+                'choices' => [
+                    'EUR' => 'EUR',
+                    'CHF' => 'CHF',
+                    'GBP' => 'GBP',
+                    'BGN' => 'BGN',
+                    'HRK' => 'HRK',
+                    'DKK' => 'DKK',
+                    'HUF' => 'HUF',
+                    'PLN' => 'PLN',
+                    'SEK' => 'SEK',
+                    'CZK' => 'CZK',
+                    'RON' => 'RON',
+                    'ALL' => 'ALL',
+                    'BYN' => 'BYN',
+                    'BAM' => 'BAM',
+                    'ISK' => 'ISK',
+                    'MKD' => 'MKD',
+                    'MDL' => 'MDL',
+                    'NOK' => 'NOK',
+                    'RSD' => 'RSD',
+                    'UAH' => 'UAH',
+                    'GIP' => 'GIP',
+                ],
+                'label' => false,
+                'placeholder' => 'currency',
+            ])
+            ->add('dueAmount', NumberType::class, [
+                'label' => 'Amount due to supplier',
+
+            ])
+            ->add('dueDate', DateType::class, [
+                'label' => 'Due date',
+                'widget' => 'single_text',
+            ])
+            ->add('exchangeRate', NumberType::class, [
+                'label' => 'Exchange rate',
+            ])
+            ->add('dueDollarsAmount', NumberType::class, [
+                'label' => 'Amount in $',
+            ])
             ->add('type', ChoiceType::class, [
                 'choices' => [
                     'Deposit' => 'deposit',
@@ -32,13 +73,16 @@ class SupplierPaymentType extends AbstractType
                 ],
                 'label' => 'Type of payment',
                 'placeholder' => 'Choose a type',
+                'required' => false,
             ])
             ->add('date', DateType::class, [
                 'widget' => 'single_text',
                 'label' => 'Date of payment',
+                'required' => false,
             ])
             ->add('paidAmount', NumberType::class, [
                 'label' => 'Payment Amount $',
+                'required' => false,
             ])
             ->add('mode', ChoiceType::class, [
                 'choices' => [
@@ -50,14 +94,16 @@ class SupplierPaymentType extends AbstractType
                 ],
                 'label' => 'Payment mode',
                 'placeholder' => 'Choose a mode',
-
+                'required' => false,
             ])
             ->add('dueCommission', NumberType::class, [
                 'label' => 'Commission due by supplier',
+                'required' => false,
             ])
             ->add('dueDateCommission', DateType::class, [
                 'widget' => 'single_text',
                 'label' => 'Commission due date',
+                'required' => false,
             ])
             ->add('status', ChoiceType::class, [
                 'choices' => [
