--- templates/accordion/supplier_information/_delete_form.html.twig
@@ -1,6 +1,0 @@
-<form method="post"
-    action="{{ path('supplier_information_delete', {'supplier_information_id': supplier_information.id, 'booking_id': supplier_information.booking.id}) }}"
-    onsubmit="return confirm('Are you sure you want to delete this item?');">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ supplier_information.id) }}">
-    <button class="btn button slide">Delete</button>
-</form>
