--- templates/accordion/supplier_information/_form.html.twig
@@ -1,4 +1,0 @@
-{{ form_start(form) }}
-    {{ form_widget(form) }}
-    <button class="btn button slide">{{ button_label|default('Save') }}</button>
-{{ form_end(form) }}
