--- templates/accordion/supplier_information/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'accordion/layout.html.twig' %}
-
-{% block title %}Edit SupplierInformation{% endblock %}
-
-{% block supplierInfo %}
-    {{ include('accordion/supplier_information/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('supplier_information_new', {'booking_id': supplier_information.booking.id}) }}">back to list</a>
-{% endblock supplierInfo %}
-{% block javascripts %}
-{{ encore_entry_script_tags('app') }}
-{{ encore_entry_script_tags('supplierInfo') }}
-{% endblock %}
