--- templates/accordion/supplier_information/new.html.twig
@@ -1,45 +1,0 @@
-{% extends 'accordion/layout.html.twig' %}
-
-{% block title %}New SupplierInformation{% endblock %}
-
-{% block supplierInfo %}
-    <div class="informations">
-        <table class="table">
-            <thead>
-                <tr>
-                    <th>Name</th>
-                    <th>Amount Due</th>
-                    <th>Amount in $</th>
-                    <th>Due Date</th>
-                    <th>actions</th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for supplier_information in supplier_informations %}
-                <tr>
-                    <td>{{ supplier_information.supplier.name }}</td>
-                    <td>{{ supplier_information.dueAmount }}</td>
-                    <td>{{ supplier_information.dueDollarsAmount|format_currency('USD', {rounding_mode: 'floor'}) }}</td>
-                    <td>{{ supplier_information.dueDate ? supplier_information.dueDate|date('m-d-Y') : '' }}</td>
-                    <td>
-                        <a href="{{ path('supplier_information_edit', {'booking_id': supplier_information.booking.id, 'supplier_information_id': supplier_information.id}) }}">edit</a>
-                        {{ include('accordion/supplier_information/_delete_form.html.twig') }}
-                    </td>
-                </tr>
-                {% else %}
-                <tr>
-                    <td colspan="8">no records found</td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-
-    {{ include('accordion/supplier_information/_form.html.twig') }}
-
-    <a href="{{ path('supplier_new', {booking_id: booking.id}) }}">Add client</a>
-{% endblock supplierInfo %}
-{% block javascripts %}
-{{ encore_entry_script_tags('app') }}
-{{ encore_entry_script_tags('supplierInfo') }}
-{% endblock %}
